improved "isRound" for type IDs
In the zettagram range of masses, some objects are round and some aren't, based on age, mass distribution and frequency of collisions. Mass distribution still can't be accounted for (SU2 doesn't provide it), but if the neighborhood is clear, collisions are unlikely making it possible for smaller objects to become round (like Mimas, Miranda, Enceladus).