more minor changes
1. Added albedo because SE calculates equilibrium temperature based on greenhouse and albeto - otherwise, SE generates albedo of a random value. In US2 there's an option to edit albedo so it's useful here too.

2. Added lum and Teff printing to all stars. For some reason this was missing from every star except the main one.

3. Removed "interior": I did the following test to see if Classifier is necessary:

interior{ breakdown... } 
or
composition{ breakdown... } 
or
interior{ composition{ breakdown... } } // like it was before

Only when "composition" is used directly, SE auto-detects helium for ice-giants vs terras and water for aquarias vs terras.
I decided to left the Classifier however, because some objects get ridiculously deep oceans and stars aren't identified well, even with this composition part sorted out.

4. Added a quick function to estimate water depth based on roughness and liquid level. If user wants ice caps and more complicated stuff though, manual edits are still needed...

A note about roughness - only objects auto-generated by US2 seem to have the default value of 21 in BumpHeight (a result of US2's ElevationToRadiusRatio). But this value can be modified from within US2 (and should because it affects terrain), and in the built-in Solar System simulations objects also have different values. Mars for example has 30~ BumpHeight because of Olympus Mons.

5. Resurrected obliquity - obliquity is really only wrong for objects orbiting binaries, and objects orbiting really close (like Mercury). In the case of mercury, I think it's because the wobble of the sun isn't accounted for in our converter... But when testing built-in Solar System simulations (and also my sims), most obliquity values are less than 3 degrees error. While this is not ideal and I think relates to inclination being wonky, it's not too bad considering that obliquity changes over time due to axial precession anyway.

Obliquity is also 100% wrong for collisional products - I haven't been able to figure out how to related the "white" axis with the "blue" axis. But those objects are rare unless that's how you tend to make objects... 

More ever, inclination, ascendingNode, argOfPericenter, and meanAnomaly are all wrong too, some with larger errors then obliquity (even when built-in US2 systems are tested). So if obliquity should be removed, maybe they should too until the math is worked out?

I haven't noticed values of -9.25596e+61 in obliquity, can you give me an example? The main object of a simulation (like Sun) does have obliquity of "6.95182e-310" which is ~0, but I'm not sure how to fix it...